# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental API",
        "description": "Dokumentacja systemu wypożyczalni",
        "version": "1.0.0"
    },
    "paths": {
...
        "/rentals/stats": {
            "get": {
                "tags": [
                    "User Rentals"
                ],
                "summary": "Pobierz statystyki lojalnościowe użytkownika",
                "operationId": "getUserRentalStats",
                "responses": {
                    "200": {
                        "description": "Zwraca liczbę wynajmów i info o nadchodzących zniżkach",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "completed_rentals": {
                                            "type": "integer"
                                        },
                                        "next_discount": {
                                            "description": "Wartość % następnej zniżki",
                                            "type": "integer"
                                        },
                                        "rentals_until_discount": {
                                            "type": "integer"
                                        },
                                        "balance": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rentals/calculate": {
            "post": {
                "tags": [
                    "User Rentals"
                ],
                "summary": "Oblicz koszt wynajmu przed złożeniem zamówienia",
                "operationId": "calculateRentalPrice",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "car_id",
                                    "rental_point_start_id",
                                    "rental_point_end_id",
                                    "start_date",
                                    "planned_end_date"
                                ],
                                "properties": {
                                    "car_id": {
                                        "type": "integer"
                                    },
                                    "rental_point_start_id": {
                                        "type": "integer"
                                    },
                                    "rental_point_end_id": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "planned_end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "use_extra_insurance": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Szczegółowa kalkulacja ceny"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rentals": {
            "post": {
                "tags": [
                    "User Rentals"
                ],
                "summary": "Stwórz nową rezerwację lub rozpocznij wynajem",
                "operationId": "storeUserRental",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "car_id": {
                                        "type": "integer"
                                    },
                                    "rental_point_start_id": {
                                        "type": "integer"
                                    },
                                    "rental_point_end_id": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "planned_end_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "use_extra_insurance": {
                                        "type": "boolean"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wynajem rozpoczęty / zarezerwowany"
                    },
                    "422": {
                        "description": "Brak środków lub auto niedostępne"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rentals/my": {
            "get": {
                "tags": [
                    "User Rentals"
                ],
                "summary": "Lista wypożyczeń zalogowanego użytkownika",
                "operationId": "getMyRentals",
                "responses": {
                    "200": {
                        "description": "Paginowana lista wynajmów"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rentals/{id}/cancel": {
            "post": {
                "tags": [
                    "User Rentals"
                ],
                "summary": "Anulowanie rezerwacji lub wczesny zwrot przez użytkownika",
                "operationId": "cancelUserRental",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Zwrócono środki na saldo"
                    },
                    "422": {
                        "description": "Nie można już anulować"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rentals/{id}/request-return": {
            "post": {
                "tags": [
                    "User Rentals"
                ],
                "summary": "Zgłoszenie zwrotu pojazdu do weryfikacji",
                "operationId": "requestReturnUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status zmieniony na pending_return"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}